refactor(ai): address PR feedback on telemetry
- Move continueAfterApproval to chatProcessing.ts (Node runtime) so
  approval-gated continuations emit OTel spans end-to-end; update the
  frontend to api.chatProcessing.continueAfterApproval.
- Rename translatedAttributes -> translateAttributes (verb form) to
  match translateSpan and project naming convention.
- Refactor attemptStream to an options object (was 6 positional args,
  convention is <= 3); derive threadId/userId from telemetry.
- Type buildTelemetryConfig return as TelemetrySettings from "ai" so
  SDK shape changes surface as type errors.
- Add edge-case tests for non-string posthog_trace_id and non-numeric
  msToFirstChunk guard branches.
- Document the module-scope OTel provider assumption.
